+++ comfyui_custom_nodes/webui_checkpoint_loader.py
@@ -1,12 +1,4 @@
-from lib_comfyui.webui_proxies import (
-    WebuiModelPatcher,
-    WebuiModelProxy,
-    WebuiClipWrapper,
-    WebuiClipProxy,
-    WebuiVaeWrapper,
-    WebuiVaeProxy,
-)
-from lib_comfyui import webui_proxies, platform_utils
+from lib_comfyui.webui import proxies
 
 
 class WebuiCheckpointLoader:
@@ -23,12 +15,12 @@ class WebuiCheckpointLoader:
     CATEGORY = "loaders"
 
     def load_checkpoint(self, void):
-        config = webui_proxies.get_comfy_model_config()
-        webui_proxies.raise_on_unsupported_model_type(config)
+        config = proxies.get_comfy_model_config()
+        proxies.raise_on_unsupported_model_type(config)
         return (
-            WebuiModelPatcher(WebuiModelProxy()),
-            WebuiClipWrapper(WebuiClipProxy()),
-            WebuiVaeWrapper(WebuiVaeProxy()),
+            proxies.ModelPatcher(proxies.Model()),
+            proxies.ClipWrapper(proxies.Clip()),
+            proxies.VaeWrapper(proxies.Vae()),
         )
