+++ comfyui_custom_nodes/webui_save_image.py
@@ -1,5 +1,5 @@
 import os
-from lib_comfyui.webui_settings import opts
+from lib_comfyui.webui.settings import opts
 from torchvision.transforms.functional import to_pil_image
